requirements.txt
kafka==1.3.5
kafka-python==2.0.2
psycopg2-binary==2.8.6
